refactor: change file organization
